+++ src/bot/commands/setup.rs
